Upload Images

+++ backend/Controllers/ProductImagesController.cs
@@ -14,10 +14,12 @@ namespace backend.Controllers
     {
         private readonly IProductImagesRepository _productImagesRepo;
         private readonly IProductRepository _productRepo;
-        public ProductImagesController(IProductImagesRepository productImagesRepo, IProductRepository productRepo)
+        private readonly IFileService _fileService;
+        public ProductImagesController(IProductImagesRepository productImagesRepo, IProductRepository productRepo, IFileService fileService)
         {
             _productImagesRepo = productImagesRepo;
             _productRepo = productRepo;
+            _fileService = fileService;
         }
 
         [HttpGet]
@@ -28,17 +30,31 @@ namespace backend.Controllers
                 return BadRequest(ModelState);
             }
             var productImages = await _productImagesRepo.GetAllAsync();
-            // var productImagesDto = productImages.Select(s => s.ToProductImagesDto());
+            
             return Ok(productImages);
         }
-        [HttpGet("{id:int}")]
-        public async Task<IActionResult> GetById([FromRoute] int id)
+        // [HttpGet("{id:int}")]
+        // public async Task<IActionResult> GetById([FromRoute] int id)
+        // {
+        //     if(!ModelState.IsValid)
+        //     {
+        //         return BadRequest(ModelState);
+        //     }
+        //     var productImages = await _productImagesRepo.GetByIdAsync(id);
+        //     if(productImages == null)
+        //     {
+        //         return NotFound();
+        //     }
+        //     return Ok(productImages);
+        // }
+        [HttpGet("{productId:int}")]
+        public async Task<IActionResult> GetByProductId([FromRoute] int productId)
         {
             if(!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
             }
-            var productImages = await _productImagesRepo.GetByIdAsync(id);
+            var productImages = await _productImagesRepo.GetByProductId(productId);
             if(productImages == null)
             {
                 return NotFound();
@@ -46,7 +62,7 @@ namespace backend.Controllers
             return Ok(productImages);
         }
         [HttpPost("{productId}")]
-        public async Task<IActionResult> Create([FromRoute] int productId, CreateProductImagesDto productImages)
+        public async Task<IActionResult> Create([FromRoute] int productId,[FromForm] CreateProductImagesDto productImages)
         {
             if(!ModelState.IsValid)
             {
@@ -56,10 +72,16 @@ namespace backend.Controllers
             {
                 return BadRequest("Product does not exist");
             }
-            var productImagesModel = productImages.ToProductImagesFromCreate(productId);
-            await _productImagesRepo.CreateAsync(productImagesModel);
+            string[]  allowedFileExtentions = [".jpg", ".jpeg", ".png"];
+            string  createdImageName = await _fileService.SaveFileAsync(productImages.Images, allowedFileExtentions);
+            var productModel = new ProductImages
+            {
+                Images = createdImageName,
+                ProductId = productId
+            };
+            await _productImagesRepo.CreateAsync(productModel);
             
-            return Ok(productImagesModel);
+            return Ok(productModel);
         }
         [HttpDelete]
         [Route("{id:int}")]
@@ -69,11 +91,21 @@ namespace backend.Controllers
             {
                 return BadRequest(ModelState);
             }
-            var productImages = await _productImagesRepo.DeleteAsync(id);
-            if(productImages == null)
+             var existingProductImage = await _productImagesRepo.GetByIdAsync(id);
+    
+            if (existingProductImage == null)
             {
-                return NotFound();
+                return NotFound(); // Trả về 404 nếu không tìm thấy hình ảnh sản phẩm
             }
+
+            var deleteResult = await _productImagesRepo.DeleteAsync(id);
+    
+            if (deleteResult == null)
+            {
+                return NotFound(); // Trả về 404 nếu việc xóa thất bại (tùy thuộc vào cách DeleteAsync hoạt động)
+            }
+
+            _fileService.DeleteFile(existingProductImage.Images);
             return Ok("Deleted product images");
         }
     }
